refactor: solve 도메인 새로 만들어서 작업하기 위해 기존 Algorithm  패키지로 원복
