+++ src/main/java/com/study/study_log/algorithm/controller/AlgorithmController.java
@@ -1,9 +1,9 @@
-package com.study.study_log.solve.controller;
+package com.study.study_log.algorithm.controller;
 
-import com.study.study_log.solve.dto.AlgorithmCreateReq;
-import com.study.study_log.solve.dto.AlgorithmUpdateReq;
-import com.study.study_log.solve.entity.Algorithm;
-import com.study.study_log.solve.service.AlgorithmService;
+import com.study.study_log.algorithm.dto.AlgorithmCreateReq;
+import com.study.study_log.algorithm.dto.AlgorithmUpdateReq;
+import com.study.study_log.algorithm.entity.Algorithm;
+import com.study.study_log.algorithm.service.AlgorithmService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
