+++ src/main/java/com/study/study_log/algorithm/repository/AlgorithmRepository.java
@@ -1,6 +1,6 @@
-package com.study.study_log.solve.repository;
+package com.study.study_log.algorithm.repository;
 
-import com.study.study_log.solve.entity.Algorithm;
+import com.study.study_log.algorithm.entity.Algorithm;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
